Added normal random number generator. Fixed compilation problems resulting from invalid char * conversions and integer conversions. Fixed a couple of misspellings in MyDXF. Fixed particle mirror test to be more tolerable to errors

+++ src/gs_solver.hpp
@@ -2,7 +2,7 @@
  *  \brief Header file for gs_solver.hpp
  */
 
-/* Copyright (c) 2005-2009 Taneli Kalvas. All rights reserved.
+/* Copyright (c) 2005-2010 Taneli Kalvas. All rights reserved.
  *
  * You can redistribute this software and/or modify it under the terms
  * of the GNU General Public License as published by the Free Software
@@ -67,7 +67,7 @@
  */
 class GSSolver : public Solver {
     double   _eps;
-    int      _imax;
+    uint32_t _imax;
     double   _w;
 
     double   _newton_Reps;  /*!< \brief Accuracy request for Newton-Raphson residual. */
@@ -82,10 +82,10 @@ public:
      */
     GSSolver( double w = 1.66,
 	      double eps = 1.0e-6, 
-	      int imax = 10000, 
+	      uint32_t imax = 10000, 
 	      double newton_Reps = 1.0e-5, 
 	      double newton_dXeps = 1.0e-6, 
-	      int newton_imax = 10 );
+	      uint32_t newton_imax = 10 );
 
     /*! \brief Destructor.
      */
@@ -106,19 +106,19 @@ public:
 
     /*! \brief Sets maximum iteration count for Newton-Raphson steps.
      */
-    void set_newton_imax( int newton_imax ) {
+    void set_newton_imax( uint32_t newton_imax ) {
 	_newton_imax = newton_imax;
     }
 
     /*! \brief Sets the accuracy request for Newton-Raphson residual.
      */
-    void set_newton_residual_eps( int newton_Reps ) {
+    void set_newton_residual_eps( double newton_Reps ) {
 	_newton_Reps = newton_Reps;
     }
 
     /*! \brief Sets the accuracy request for Newton-Raphson step size.
      */
-    void set_newton_step_eps( int newton_dXeps ) {
+    void set_newton_step_eps( double newton_dXeps ) {
 	_newton_dXeps = newton_dXeps;
     }
 
@@ -129,7 +129,7 @@ public:
      *  library.
      */
     static bool gauss_seidel( const CRowMatrix &mat, const Vector &rhs, Vector &sol,
-			      int &imax, double &eps, double w );
+			      uint32_t &imax, double &eps, double w );
 };
